% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVXfromR-package.R
\docType{package}
\name{CVXfromR-package}
\alias{CVXfromR-package}
\alias{CVXfromR}
\title{Matlab and CVX from R}
\description{
Call Matlab and, in particular, CVX without leaving R.
}
\details{
This package contains three functions: \code{\link{CallMatlab}},
\code{\link{CallCVX}}, and \code{\link{CallCVX.varyparam}}.

Note: MATLAB and CVX are not included.  Assumes that MATLAB can be opened
from the command line (see \code{matlab.code} argument for details).
}
\author{
Jacob Bien

Maintainer: Jacob Bien <jbien@cornell.edu>
}
